Support version-10 draft

+++ inc/teep/teep_message_print.h
@@ -16,7 +16,7 @@
 
 const char *teep_err_to_str(teep_err_t err);
 const char *teep_err_code_to_str(int32_t err_code);
-const char *teep_cose_algs_key_to_str(int32_t cose_algs_key);
+const char *teep_cose_algs_key_to_str(int64_t cose_algs_key);
 teep_err_t teep_print_query_request(const teep_query_request_t *query_request, uint32_t indent_space);
 teep_err_t teep_print_query_response(const teep_query_response_t *query_response, uint32_t indent_space);
 teep_err_t teep_print_update(const teep_update_t *teep_update, uint32_t indent_space, const char *ta_public_key);
